feat: reuse inline-chunk plugin in builder
