+++ packages/builder/builder-shared/src/plugins/index.ts
@@ -4,3 +4,4 @@ export type { HtmlTagsPluginOptions } from './HtmlTagsPlugin';
 export { HtmlCrossOriginPlugin } from './HtmlCrossOriginPlugin';
 export { HtmlAppIconPlugin } from './HtmlAppIconPlugin';
 export { HtmlFaviconUrlPlugin, type FaviconUrls } from './HtmlFaviconUrlPlugin';
+export { InlineChunkHtmlPlugin } from './InlineChunkHtmlPlugin';
