+++ packages/builder/builder-webpack-provider/tests/plugins/inlineChunk.test.ts
@@ -1,7 +1,7 @@
 import { expect, describe, it, beforeEach, afterEach } from 'vitest';
 import { builderPluginEntry } from '@builder/plugins/entry';
 import { builderPluginHtml } from '@builder/plugins/html';
-import { builderPluginInlineChunk } from '@/plugins/inlineChunk';
+import { builderPluginInlineChunk } from '@builder/plugins/inlineChunk';
 import { createStubBuilder } from '@/stub';
 
 describe('plugins/inlineChunk', () => {
